+++ RELEASE_NOTES.md
@@ -1,3 +1,8 @@
+### 3.12.0 - 2020-08-29
+* Parameter nullability inference for parsable queries
+* Detecting the missing columns which are required for INSERT queries
+* Better error messages when reading from a result set which doesn't return any columns
+
 ### 3.11.0 - 2020-08-18
 * Even better error messages that include whether types were arrays or not